Now calling API from last published
