+++ main.py
@@ -68,7 +68,8 @@ is_first_time_setup = tweeting.create_db(db_filename, create_str)
 periodicity_s = 3600
 max_time = 7*24*3600
 
-thread = tweeting.RepeatEvery(periodicity_s, tweeting.tweet_news, tweepyapi, apiKey, qaly_path, error_log_filename, error_log_pointer, db_filename, is_first_time_setup)
+thread = tweeting.RepeatEvery(periodicity_s, tweeting.tweet_news, tweepyapi, apiKey, qaly_path, error_log_filename, error_log_pointer, db_filename, is_first_time_setup,
+dbg_mode = True)
 
 print('Starting')
 thread.start()
